+++ MacPacker/CHANGELOG.md
@@ -5,6 +5,7 @@ v0.11
 - fix: Quick look extension missing in App Store version
 - fix: Empty window opened in addition to file opened with 'Open With' option in Finder
 - core: Re-enable macOS 13 as minimum deployment target
+- core: Language support for Italian
 
 v0.10
 - feat: Quick Look extension